pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mohtion"
version = "0.1.0"
description = "Autonomous Tech Debt Bounty Hunter - A GitHub App that finds and fixes technical debt"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Mohtion Team" }
]
keywords = ["ai", "refactoring", "technical-debt", "github-app", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "arq>=0.25.0",
    "httpx>=0.26.0",
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "pyjwt[crypto]>=2.8.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
mohtion = "mohtion.cli:main"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
